rename seccomp package to seccompiler
While the binary was named seccompiler, the library crate was
still named `seccomp`.

This is helpful, in order to distinguish between the old
library and the new, seccompiler library.

This also makes sense and comes in handy in the eventuality
of moving the code to a standalone repo.
