+++ src/seccompiler/src/compiler.rs
