+++ src/vmm/Cargo.toml
@@ -21,7 +21,7 @@ logger = { path = "../logger" }
 mmds = { path = "../mmds" }
 polly = { path = "../polly" }
 rate_limiter = { path = "../rate_limiter" }
-seccomp = { path = "../seccomp" }
+seccompiler = { path = "../seccompiler" }
 snapshot = { path = "../snapshot"}
 utils = { path = "../utils" }
