+++ src/vmm/src/builder.rs
@@ -30,7 +30,7 @@ use devices::virtio::{Balloon, Block, MmioTransport, Net, VirtioDevice, Vsock, V
 use kernel::cmdline::Cmdline as KernelCmdline;
 use logger::{error, warn};
 use polly::event_manager::{Error as EventManagerError, EventManager, Subscriber};
-use seccomp::BpfThreadMap;
+use seccompiler::BpfThreadMap;
 use snapshot::Persist;
 use utils::eventfd::EventFd;
 use utils::terminal::Terminal;
@@ -375,7 +375,7 @@ pub fn build_microvm_for_boot(
     // Execution panics if filters cannot be loaded, use --no-seccomp if skipping filters
     // altogether is the desired behaviour.
     // Keep this as the last step before resuming vcpus.
-    seccomp::apply_filter(
+    seccompiler::apply_filter(
         &seccomp_filters
             .remove("vmm")
             .ok_or_else(|| MissingSeccompFilters("vmm".to_string()))?,
@@ -466,7 +466,7 @@ pub fn build_microvm_from_snapshot(
 
     // Load seccomp filters for the VMM thread.
     // Keep this as the last step of the building process.
-    seccomp::apply_filter(
+    seccompiler::apply_filter(
         &seccomp_filters
             .remove("vmm")
             .ok_or_else(|| MissingSeccompFilters("vmm".to_string()))?,
