+++ src/vmm/src/vstate/vcpu/mod.rs
@@ -27,7 +27,7 @@ use crate::{
 use kvm_bindings::{KVM_SYSTEM_EVENT_RESET, KVM_SYSTEM_EVENT_SHUTDOWN};
 use kvm_ioctls::VcpuExit;
 use logger::{error, info, IncMetric, METRICS};
-use seccomp::{BpfProgram, BpfProgramRef};
+use seccompiler::{BpfProgram, BpfProgramRef};
 use utils::{
     errno,
     eventfd::EventFd,
@@ -268,7 +268,7 @@ impl Vcpu {
         // Load seccomp filters for this vCPU thread.
         // Execution panics if filters cannot be loaded, use --no-seccomp if skipping filters
         // altogether is the desired behaviour.
-        if let Err(e) = seccomp::apply_filter(seccomp_filter) {
+        if let Err(e) = seccompiler::apply_filter(seccomp_filter) {
             panic!(
                 "Failed to set the requested seccomp filters on vCPU {}: Error: {}",
                 self.kvm_vcpu.index, e
