+++ tools/devtool
@@ -687,7 +687,7 @@ cmd_build() {
         --workdir "$CTR_FC_ROOT_DIR" \
         ${extra_args} \
         -- \
-        cargo build -p seccomp --bin seccompiler \
+        cargo build -p seccompiler --bin seccompiler \
             --target-dir "$CTR_CARGO_SECCOMPILER_TARGET_DIR" \
             "${cargo_args[@]}"
     ret=$?
